+++ lang/sl/email.php
@@ -57,7 +57,8 @@ return [
     'personal_access_token_details' => 'Podatki žetona',
     'personal_access_token_name' => 'Ime žetona',
     'personal_access_token_valid_until' => 'Veljaven do',
-    'if_this_was_not_you_personal_access_token' => 'V primeru, da tega niste naredili vi, prosimo nemudoma izbrišite žeton in kontaktirajte administratorja.',
+    'if_this_was_not_you_personal_access_token' =>
+        'V primeru, da tega niste naredili vi, prosimo nemudoma izbrišite žeton in kontaktirajte administratorja.',
     'subject_support_ticket' => 'Obvestilo o novem podpornem zahtevku',
     'new_support_ticket' => 'Za vas je bil ustvarjen nov podporni zahtevek. ',
     'support_ticket_details' => 'Podrobnosti podpornega zahtevka',