+++ resources/js/mixins/admin/status.js
@@ -2,6 +2,7 @@ export default {
     data() {
         return {
             version_info: [],
+            status_info: {},
         }
     },
 
@@ -13,6 +14,16 @@ export default {
         getVersion() {
             this.makeHttpRequest('GET', '/api/admin/version').then((response) => {
                 this.version_info = response.data.data
+            })
+        },
+
+        /**
+         * Get status of the application
+         * @return {void} return status
+         */
+        getStatus() {
+            this.makeHttpRequest('GET', '/api/admin/status').then((response) => {
+                this.status_info = response.data.data
             })
         },
     },