+++ resources/js/views/admin/settings/StatusPage.vue
@@ -12,10 +12,39 @@
                     </Button>
                 </div>
             </Message>
-            <div id="status_page" class="card">
+            <div id="version_data_card" class="card">
                 <DisplayData :value="version_info?.current_version" :input="$t('admin.status.app_version')" />
 
                 <DisplayData :value="version_info?.latest_version" :input="$t('admin.status.latest_version')" />
+            </div>
+
+            <div id="storage_used_card" class="card mt-4">
+                <DisplayData :value="status_info?.php_version" :input="$t('admin.status.php_version')" />
+
+                <DisplayData :value="status_info?.server_software" :input="$t('admin.status.server_software')" />
+
+                <DisplayData :value="status_info?.database_connection" :input="$t('admin.status.database_connection')" />
+
+                <DisplayData :value="status_info?.database_version" :input="$t('admin.status.database_version')" />
+
+                <DisplayData
+                    :value="status_info?.storage_limit == -1 ? $t('admin.status.unlimited') : formatFileSize(status_info?.storage_limit)"
+                    :input="$t('admin.status.storage_limit')"
+                />
+
+                <DisplayData :value="formatFileSize(status_info?.storage_used)" :input="$t('admin.status.storage_used')" />
+
+                <DisplayData
+                    v-if="status_info?.storage_limit != -1"
+                    :value="formatFileSize(status_info?.storage_free)"
+                    :input="$t('admin.status.storage_free')"
+                />
+
+                <DisplayData
+                    v-if="status_info?.storage_limit != -1"
+                    :value="status_info?.storage_usage_percentage + '%'"
+                    :input="$t('admin.status.storage_usage_percentage')"
+                />
             </div>
         </LoadingScreen>
 
@@ -40,6 +69,7 @@ export default {
 
     created() {
         this.getVersion()
+        this.getStatus()
     },
     methods: {
         showChangelogDialog() {